Refactor modules into separate scripts
